requirements.txt
aniso8601==6.0.0
APScheduler==3.6.0
bcrypt==3.1.6
certifi==2019.3.9
chardet==3.0.4
clarifai==2.6.1
configparser==3.7.4
Click==7.0
cssselect==1.0.3
cycler==0.10.0
dnspython==1.16.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==3.18.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.7
future==0.17.1
googleapis-common-protos==1.5.10
gunicorn==19.9.0
idna==2.8
isodate==0.6.0
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.3
passlib==1.7.1
protobuf==3.7.1
pycparser==2.19
PyJWT==1.7.1
pymongo==3.8.0
pyparsing==2.4.0
pyquery==1.4.0
PySocks==1.7.0
python-dateutil==2.7.3
pytz==2018.7
requests==2.21.0
scipy==1.2.1
six==1.12.0
twilio==6.25.2
urllib3==1.24.2
virtualenv==16.4.1
websocket-client==0.54.0